backend_server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7,<3.0
python-multipart==0.0.6
yfinance
pandas>=2.2.0
numpy>=1.26.0
alpha-vantage>=3.0.0
pydantic-settings>=2.2.0
SQLAlchemy>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
google-auth>=2.17.0
requests>=2.31.0
python-dotenv>=1.0.1
